Renamed SearchSelect to EntitySearch, and added icons to distinguish entity types.
