+++ src/client/components/forms/parts/editionData.jsx
@@ -20,7 +20,7 @@
 var React = require('react');
 var PartialDate = require('../../input/partialDate.jsx');
 var Select = require('../../input/select2.jsx');
-var SearchSelect = require('../../input/search-select.jsx');
+var SearchSelect = require('../../input/entity-search.jsx');
 var Input = require('react-bootstrap').Input;
 var Button = require('react-bootstrap').Button;
 var Identifiers = require('./identifiers.jsx');
